ga_metrics.bulk, shifts most of logic into ga_metrics.core. settings.py, removes notion of environments (DEV, PROD, etc)
