+++ elife.cfg
@@ -1,6 +1,6 @@
 [general]
 debug: True
-env: dev
+log-file:
 secret-key: these-are-dev-settings.DO.NOT.USE.IN.PROD.EVER
 allowed-hosts:
